Moved TopologyUtils from spi to common
