+++ heron/packing/src/java/BUILD
@@ -5,6 +5,7 @@ load("/tools/rules/heron_deps", "heron_java_proto_files")
 packing_deps_files = [
     "//heron/spi/src/java:common-spi-java",
     "//heron/common/src/java:basics-java",
+    "//heron/common/src/java:utils-java",
     "//heron/spi/src/java:packing-spi-java",
     "@com_google_guava_guava//jar",
     ":builder",
@@ -32,6 +33,7 @@ java_library(
     srcs = glob(["**/utils/*.java"]),
     deps = heron_java_proto_files() + [
         "//heron/common/src/java:basics-java",
+        "//heron/common/src/java:utils-java",
         "//heron/spi/src/java:common-spi-java",
         "//heron/spi/src/java:packing-spi-java",
         "//heron/spi/src/java:utils-spi-java",
