+++ heron/spi/src/java/BUILD
@@ -30,6 +30,7 @@ java_library(
         "//heron/api/src/java:classification",
         "//heron/common/src/java:basics-java",
         "//heron/common/src/java:config-java",
+        "//heron/common/src/java:utils-java",
         "//heron/proto:proto_common_java",
         "//heron/proto:proto_execution_state_java",
         "//heron/proto:proto_scheduler_java",
@@ -54,6 +55,7 @@ utils_deps_files = \
         ":metricsmgr-spi-java",
         "//heron/common/src/java:basics-java",
         "//heron/common/src/java:config-java",
+        "//heron/common/src/java:utils-java",
         "//heron/api/src/java:api-java",
         "@com_google_guava_guava//jar",
     ]
