+++ heron/tools/apiserver/src/java/com/twitter/heron/apiserver/actions/ActionFactoryImpl.java
@@ -15,8 +15,8 @@ package com.twitter.heron.apiserver.actions;
 
 import com.twitter.heron.api.generated.TopologyAPI;
 import com.twitter.heron.apiserver.utils.ConfigUtils;
+import com.twitter.heron.common.utils.topology.TopologyUtils;
 import com.twitter.heron.spi.common.Config;
-import com.twitter.heron.spi.utils.TopologyUtils;
 
 public class ActionFactoryImpl implements ActionFactory {
